Add docopt and dateutil as dependencies
docopt is added to generate help messages and dateutil is added to parse
the dates in the news feed into date objects.

This also adds the pacman hook. The hook is currently untested and
depending on the install location of informant, will need to be updated
to properly exec it.