Add basic functionality
This gets the subcommands check and read working mostly as they should
from an output standpoint. They need to actually save the read items
into a file somewhere (probably under /var/, but should probably be
configurable).

Also the newly added '--prompt' argument for the check command needs to
be used to display a prompt asking the user if they would like to
continue their pacman transaction in spite of unread news items.