Implement '--prompt' argument for 'check' command
This is the last 'core' functionality aside from persistence of read
items. This is now in acceptable state to actually be installed, yet I'm
leaving the version number in alpha state because some things just plain
don't work yet.

This also adds a debug option which prints the arguments as parsed by
docopt, a file option (so far unused) and a license.