Rework check and add list commands
This changes how the check command will work. Since pacman doesn't allow
interactive hooks the way the hook will work now is if there is only
one unread item it will display it and mark as read, but still exiting
with failure in case the news item requires user intervention. If there
are multiple unread items it will print a message saying so and return
with failure forcing the user to read them with informant.

This also adds a list command which will print the most recent items
regardless of read status.