Add file error checking
This adds much more error checking to the initial file I/O. Getting the
readlist was extracted out into it's own function to minimize repeated
code between has_been_read and mark_as_read. Currently the file is being
read every time either of these functions gets called, which is fine for
the 'read' subcommand (as it only operates on one feed item) but for the
'check' subcommand there is a lot of unnecessary file I/O going on.

This also creates a function for printing feed items so that some
additional processing can be done on them. Currently only paragraph tags
are stripped.

Finally, the 'list' subcommand has been updated to also print indexes
next to titles.