Make readlist a global variable
This makes the readlist a global variable, and reads from the disk once
right after parsing the arguments. This is a significant improvement on
the number of reads done from the disk, previously every time
has_been_read was called informant would read from the disk.

This also changes the signatures for has_been_read and mark_as_read.
Previously they both took a title and a datetime, but now they both just
take an entry. This makes it a lot cleaner to call the functions.