Add color printing for error
This adds color to the error print statement to draw more user attention
to it. It also adds an error status code exit to the error so that the
user's shell may be able to draw attention to the fact that something
went wrong.

Color printing is done using ANSI escape strings, this seems like it
will be the easiest and most supported way of printing in color.

Currently there is only one error message where this needed to be used,
but this should be the pattern for any future error messages. Also, I
am considering disabling the ANSI codes in 'raw' mode, maybe a separate
'no-color' option would be better.