Unpack tuple in get_datfile to handle new datfile
The failure caused by the new datfile format is a ValueError due to
unpacking the tuple. This unpacks the tuple in the get_datfile()
function so that this case can be checked for and handled gracefully.

The newest version of informant will be accompanied by a notice about
the new datfile format.