Add pacman detection and better messages
Fixes github.com/bradford-smith94/informant/issues/22

Related github.com/bradford-smith94/informant/issues/20

This makes use of os.getppid() and `ps` to detect if the script is being
run from pacman. This detection allows us to format messages differently
when running as the pacman hook vs when running interactively.

The pacman detection also allows for not blocking pacman upgrades when
informant fails (such as with a lack of internet connection).