Add config file to specify feeds
This adds a config file that allows you to specify different and/or
multiple feeds for informant to read from. Overall structure was updated
so that the feeds are processed into one uniform feed with matching
properties.

Since each feed can have different property names the config file has
lines to specify key names so that the feed is parsed correctly.

Currently this will break the feed caching. Perhaps we should cache the
component feeds (with different cache timeouts?). The Archlinux32 news
feed doesn't have a cache max-age so it would need to be informant's
decision or fetched every time.

The config file also needs to be made optional so that by default the
Arch news feed will be the only one processed.