Setup partial caching and feed building
This sets up a structure of "component feeds" which are the list of all
feeds the user wants informant to read. Since these are each separate
feeds they can have their own notions of how long they are valid in the
cache, which presents us with a partial cache validity issue.

To solve the issue of partial caching this introduces ways to build the
informant "feed" (combined feed of everything the user wants) from each
of it's components. The build_feed() function allow us to specify which
sets of components are already valid and which need to be requested.

The write_cache() and cache_is_valid() functions received significant
updates to account for the new structure. One of the test feeds
(Archlinux32 news feed) does not have a 'max-age' value in it's
'Cache-Control' header, so this also adds handling for that, falling
back to using the 'Expires' header.