Resolve globals issue
This makes sure the main() function is set to update global values so
that minimal changes are needed to the rest of the code structure.

Also it was found that in order for CacheControl to use a FileCache the
module 'lockfile' also needs to be installed, so this updates the setup
file as well.