Lowercase all feed headers before checking
Related: https://github.com/bradford-smith94/informant/pull/25

0f037d575368c21b8e3c872cb849513c0904e598 changed the key informant
checks to be lowercase which matches the output of newer versions of
python-feedparser (known 6.0.2) but not older versions (known 5.2.1).

This adds an additional level of safety by overwriting the feed headers
to be lowercase before checking for that key. The headers should be
case-insensitive according to spec anyway [1].

[1]: https://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4.2