Fix permissions of cache files
This adds permissions to the FileCache object, in order for them to be
set correctly we need to make sure that we have a less restrictive umask
active (Arch default umask seems to be 0o0022).

After unrestricting the umask adding the options to FileCache will save
files with the correct permissions but incorrect group. This is resolved
by setting the setgid bit on /var/cache/informant so that any files
created in that directory are owned by it's gid (which will be the
informant group) [1]. This change needs to be made in the PKGBUILD for
release.

[1]: https://www.liquidweb.com/kb/how-do-i-set-up-setuid-setgid-and-sticky-bits-on-linux/
(see section called "Setgid on Directories")