Do not block pacman on fetch error
Closes https://github.com/bradford-smith94/informant/issues/36

In the case that there is no internet connection informant previously
exited with error status (because no feed existed to act on). This
caused the pacman hook to always fail on install/upgrade transactions
which could prevent the user from fixing an internet connection issue.

This checks the type of error we catch from CacheControl/FeedParser to
see if the error was likely a connection error (by way of checking for a
urllib.URLError) and printing a warning message if this is the case.
Even when the error is not a connection issue the informant.feed class
will only return an empty feed (in case there are multiple feeds and
only one fails) and will print a separate warning if across all feeds
there are no items to act on.

This also leaves room for an unexpected error (if FeedParser changes how
feedparser.util.FeedParserDict works -- this was used to create the
empty feed), but even this unexpected error will not block pacman.