Update default archnews url
Fix https://github.com/bradford-smith94/informant/issues/43

When issue 43 was reported it seems that the Archlinux news URL now has
a 301 redirect to the same thing with a trailing slash. This fixes the
issue by defaulting to the trailing slash but a further update should be
done at some point to check for and possibly follow redirects.